started to implement udp example

+++ forwarder/logging/MqttSnForwarderLoggingMessages.c
@@ -4,7 +4,6 @@
 
 #include "MqttSnForwarderLoggingMessages.h"
 #include "MqttSnForwarderLoggingBasic.h"
-#include <netinet/in.h>
 
 #define MQTT_SN_MESSAGE_TYPE_STRING_NEGATIVE_ENUMS_OFFSET 2
 static const char *MQTT_SN_MESSAGE_TYPE_STRING[] = {
@@ -48,7 +47,7 @@ int log_radius(const MqttSnLogger *logger, uint8_t radius) {
 int log_gw_add(const MqttSnLogger *logger, const device_address *gwAdd, uint16_t gwAdd_len) {
   log_str(logger, "gwAdd");
   const char *dot = ".";
-  for (size_t i = 0; i < gwAdd_len; ++i) {
+  for (uint16_t i = 0; i < gwAdd_len; ++i) {
     if (log_uint8(logger, gwAdd->bytes[i])) {
       return -1;
     }
